% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/crafter-package.r
\docType{package}
\name{crafter}
\alias{crafter}
\alias{crafter-package}
\title{A package to work with packet captures}
\description{
A package to work with packet captures
}
\author{
Bob Rudis (@hrbrmstr)
}

