% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/crafter.r
\name{read_pcap}
\alias{read_pcap}
\title{Read in a packet capture file}
\usage{
read_pcap(capture_file, filter = "")
}
\arguments{
\item{capture_file}{path to a PCAP}

\item{filter}{filter}
}
\description{
Reads in a packet capture file and returns an hyrid S3 object with internal
methods to call to operate on the capture.
}
\references{
\href{http://www.tcpdump.org/manpages/pcap-filter.7.html}{pcap filter reference}
}

