% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_in.R
\name{seq_in}
\alias{seq_in}
\title{Find a (the first) sequence in a vector}
\usage{
seq_in(source_vector, pattern_vector)
}
\arguments{
\item{source_vector}{vector to search in}

\item{pattern_vector}{vector that has the sequence to search for}
}
\description{
Find a (the first) sequence in a vector
}
\examples{
seq_in(c(1,5,4,3,2,1,5), c(4,3, 2))
}
