% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/crafter.r
\name{read_pcap}
\alias{read_pcap}
\title{Read in a packet capture file}
\usage{
read_pcap(capture_file, filter = "")
}
\arguments{
\item{capture_file}{path to a PCAP}

\item{filter}{filter}
}
\value{
A \code{crafter} hybrid S3 object
}
\description{
Reads in a packet capture file and returns an hyrid S3 \code{crafter} object
with internal methods and accessible variables.
}
\details{
Callable internal methods & accesible variables  of the \code{crafter} object:

\describe{
  \item{\code{packet_count}}{Total packets in capture}
  \item{\code{get_ips(src_or_dst)}}{Retrieve a data frame of IP addresses.
        "\code{src_or_dst}" can be one of "\code{src}", "\code{dst}" or "\code{all}" }
  \item{\code{packet_info()}}{Retrieve a data frame of high-level packet information}
  \item{\code{get_payload(num)}}{Retrieve the payload of a specific packet number}
  \item{\code{get_layer(layer_name)}}{Get all the packets from a layer,
       "\code{layer_name}" can be one of \code{TCP, IP, ICMP}}
  \item{\code{summary()}}{Prints summary information about the packet capture}
}
}
\examples{
ncap <- read_pcap(system.file("pcaps/Ncapture.pcap", package="crafter"))

# get overall capture summary
summary(ncap)

# same thing
ncap$summary()

# high level packet info
ncap_info <- ncap$packet_info()
head(ncap_info)

# retrieve specific layers
ncap_ip <- ncap$get_layer("IP")
head(ncap_ip)

ncap_tcp <- ncap$get_layer("TCP")
head(ncap_tcp)

# get the IPs from the capture
ncap$get_ips("all")

# get packet payload
ncap$get_payload(7119)
}
\references{
\href{http://www.tcpdump.org/manpages/pcap-filter.7.html}{pcap filter reference}
}

